IOFireWireSBP2-442
Imported from IOFireWireSBP2-442.tar.gz

+++ Source/IOFireWireSBP2/IOFireWireSBP2ManagementORB.cpp
@@ -451,7 +451,7 @@ IOReturn IOFireWireSBP2ManagementORB::setResponseBuffer
     
 	if( desc != NULL )
     {
-        len = desc->getLength();
+        len = (UInt32)desc->getLength();
 		
 		fResponseMap = desc->map();
 		if( fResponseMap != NULL )
